Moved AudioPlayer and NftCard from UI lib to main project
