+++ components/items/ItemsGrid/ItemsGrid.vue
@@ -40,7 +40,7 @@
 
       <!-- skeleton on fetching next page -->
       <template v-if="isLoading || isFetchingData">
-        <NeoNftCardSkeleton
+        <NftCardSkeleton
           v-for="n in skeletonCount"
           :key="n"
           :hide-media-info="hideMediaInfo" />
@@ -54,7 +54,7 @@
     <DynamicGrid
       v-if="total === 0 && (isLoading || isFetchingData)"
       class="my-5">
-      <NeoNftCardSkeleton
+      <NftCardSkeleton
         v-for="n in skeletonCount"
         :key="n"
         :hide-media-info="hideMediaInfo" />
@@ -66,7 +66,6 @@
 </template>
 
 <script setup lang="ts">
-import { NeoNftCardSkeleton } from '@kodadot1/brick'
 import DynamicGrid from '@/components/shared/DynamicGrid.vue'
 import ItemsGridImage from './ItemsGridImage.vue'
 import ItemsGridImageTokenEntity from './ItemsGridImageTokenEntity.vue'
