+++ components/shared/audioPlayer/AudioPlayer.vue
@@ -34,7 +34,7 @@
           'is-flex is-align-items-center is-justify-items-center':
             !isWaveformReady,
         }">
-        <NeoWaveform
+        <Waveform
           v-show="isWaveformReady"
           :class="{
             'is-clickable': canStartPlaying,
@@ -64,7 +64,7 @@ import { computed, ref } from 'vue'
 import { useEventListener, useMediaControls } from '@vueuse/core'
 import { NeoButton, NeoIcon, NeoSkeleton } from '@kodadot1/brick'
 import { getRandomValues } from '@/components/unique/utils'
-import NeoWaveform from './NeoWaveform.vue'
+import Waveform from './Waveform.vue'
 
 defineProps<{
   src?: string
@@ -121,7 +121,7 @@ const pushToActionStack = ({ promise, reject }) => {
 const play = (time?: number) => {
   return new Promise((resolve, reject) => {
     if (!canPlay.value) {
-      return reject(new Error("Player: Can't play"))
+      return reject(new Error("Player: Can't play")) // eslint-disable-line quotes
     }
     flushPreviouseActions()
       .then(() => {
@@ -152,7 +152,7 @@ const play = (time?: number) => {
 const pause = async () => {
   return new Promise((resolve, reject) => {
     if (!canPause.value) {
-      return reject(new Error("Player: Can't pause"))
+      return reject(new Error("Player: Can't pause")) // eslint-disable-line quotes
     }
 
     flushPreviouseActions()
