+++ components/shared/nftCard/NftMediaInfoStacked.vue
@@ -62,9 +62,9 @@
 </template>
 
 <script lang="ts" setup>
-import type { NeoNFT } from './types'
+import type { NeoNFT, NftCardVariant } from './types'
 import { computed } from 'vue'
-import { NeoButton, type NftCardVariant } from '@kodadot1/brick'
+import { NeoButton } from '@kodadot1/brick'
 
 import CommonTokenMoney from '@/components/shared/CommonTokenMoney.vue'
 
@@ -94,7 +94,7 @@ const collectionNameLabel = computed(() => props.token.collection.name || '--')
 </script>
 
 <style lang="scss" scoped>
-@import './NeoNftCard.scss';
+@import './NftCard.scss';
 
 .staked-primary-title {
   margin-top: 6px;
