+++ components/shared/nftCard/types.ts
@@ -14,3 +14,9 @@ export type NeoNFT = {
     name: string
   }
 }
+
+export type NftCardVariant =
+  | 'primary'
+  | 'minimal'
+  | 'stacked-primary'
+  | 'stacked-minimal'
