+++ libs/ui/src/types.ts
@@ -15,9 +15,3 @@ export type NeoButtonVariant =
   | 'text'
   | 'pill'
   | 'border-icon'
-
-export type NftCardVariant =
-  | 'primary'
-  | 'minimal'
-  | 'stacked-primary'
-  | 'stacked-minimal'
